+++ src/generator/painter/circle.rs
@@ -26,7 +26,7 @@ pub struct Options {
 	pub radius_bias: f64, // 0 = normal; -1 = quad bias towards small, 1 = quad bias towards big, etc
 	pub anti_alias: bool,
 	pub color_seed: f64,
-	pub rng_seed: u128,
+	pub rng_seed: u32,
 	pub margins: Margins<SizeUnit>,
 }
 
@@ -45,7 +45,7 @@ impl CirclePainter {
 			radius_bias: 0.0f64,
 			anti_alias: true,
 			color_seed: 0.0f64,
-			rng_seed: 0u128,
+			rng_seed: 0,
 			margins: Margins::<SizeUnit> {
 				top: SizeUnit::Pixels(0),
 				right: SizeUnit::Pixels(0),
@@ -61,7 +61,7 @@ impl CirclePainter {
 }
 
 impl Painter for CirclePainter {
-	fn paint(&self, canvas: &RgbImage, iteration: u64, seed_map: &RgbImage) -> Result<RgbImage, &str> {
+	fn paint(&self, canvas: &RgbImage, iteration: u32, seed_map: &RgbImage) -> Result<RgbImage, &str> {
 		let mut rng = get_rng(self.options.rng_seed, iteration);
 
 		let image_area = canvas.dimensions();
