+++ src/generator/painter/rect.rs
@@ -27,7 +27,7 @@ pub struct Options {
 	pub width_bias: f64, // 0 = normal; -1 = quad bias towards small, 1 = quad bias towards big, etc
 	pub height_bias: f64, // 0 = normal; -1 = quad bias towards small, 1 = quad bias towards big, etc
 	pub color_seed: f64,
-	pub rng_seed: u128,
+	pub rng_seed: u32,
 	pub margins: Margins<SizeUnit>,
 }
 
@@ -50,7 +50,7 @@ impl RectPainter {
 			}],
 			height_bias: 0.0f64,
 			color_seed: 0.0f64,
-			rng_seed: 0u128,
+			rng_seed: 0,
 			margins: Margins::<SizeUnit> {
 				top: SizeUnit::Pixels(0),
 				right: SizeUnit::Pixels(0),
@@ -66,7 +66,7 @@ impl RectPainter {
 }
 
 impl Painter for RectPainter {
-	fn paint(&self, canvas: &RgbImage, iteration: u64, seed_map: &RgbImage) -> Result<RgbImage, &str> {
+	fn paint(&self, canvas: &RgbImage, iteration: u32, seed_map: &RgbImage) -> Result<RgbImage, &str> {
 		let mut rng = get_rng(self.options.rng_seed, iteration);
 
 		let image_area = canvas.dimensions();
